add stock market solution
